Delegated activities to separate Kotlin activity files.
