+++ app/build.gradle
@@ -57,7 +57,6 @@ dependencies {
     implementation 'androidx.compose.ui:ui-tooling-preview'
     implementation 'androidx.compose.material3:material3'
     implementation 'com.google.firebase:firebase-auth:22.0.0'
-    implementation 'androidx.core:core-ktx:+'
 
     testImplementation 'junit:junit:4.13.2'
